Refactored range record
